requirements.txt
streamlit==1.32.0
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
wikipedia==1.4.0
markdown==3.5.1
